backend/requirements.txt
# Core Django
Django==5.2.6
djangorestframework==3.15.2

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Production Server
gunicorn==23.0.0

# Static Files
whitenoise==6.8.2

# CORS
django-cors-headers==4.6.0

# Environment
python-dotenv==1.0.1

# HTTP Requests
requests==2.32.3

# Security (Production)
django-csp==3.8
django-ratelimit==4.1.0